rustinfo/requirements.txt
aiohttp>=3.10.0
aiosignal>=1.3.0
aiosqlite>=0.19.0
attrs>=25.0.0
certifi>=2025.0.0
charset-normalizer>=3.4.0
colorama>=0.4.6
discord.py>=2.3.0
frozenlist>=1.5.0
idna>=3.10
multidict>=6.1.0
propcache>=0.3.0
python-dotenv>=1.0.0
requests>=2.32.0
rustplus>=2.0.0
urllib3>=2.3.0
yarl>=1.18.0
pillow>=10.0.0